Impl: Some obvious pruning.
It's impossible to make two disjoint subsets that share a sum that is greater than half the of sum of the original.
We can break out of loops whenever we encounter a sum that's greater than half of the sum of the original.

Benchmarking shows that this is actually a much more significant impovement than I imagined.